Move Interfaces into class files
