+++ src/App.ts
@@ -1,5 +1,9 @@
-import IMain from './IMain';
-import IApp from './IApp';
+import { IMain } from './Main';
+
+export interface IApp {
+  main: IMain;
+  start(): void;
+}
 
 export default class App implements IApp {
   main: IMain;
