+++ src/CLI.ts
@@ -1,7 +1,12 @@
-import ICLI from './ICLI';
 import { Answers, Question } from 'inquirer';
 const clear = require('clear');
 const inquirer = require('inquirer');
+
+export interface ICLI {
+  print(message: string): void;
+  clear(): void;
+  prompt(questions: Question[]): Promise<Answers>;
+}
 
 export default class CLI implements ICLI {
   constructor(){}
