--- src/IApp.d.ts
@@ -1,6 +1,0 @@
-import IMain from "./IMain";
-
-export default interface IApp {
-  main: IMain;
-  start(): void;
-}
