--- src/ICLI.d.ts
@@ -1,7 +1,0 @@
-import { Answers, Question } from "inquirer";
-
-export default interface ICLI {
-  print(message: string): void;
-  clear(): void;
-  prompt(questions: Question[]): Promise<Answers>;
-}
