--- src/IMain.d.ts
@@ -1,3 +1,0 @@
-export default interface IMain {
-  start(): Promise<void>;
-}
